Workaround to get round trip, distance a bit unprecise

+++ src/main/java/com/api/route_calculator/service/RouteService.java
@@ -5,5 +5,5 @@ import com.api.route_calculator.entity.route.Route;
 import java.util.ArrayList;
 
 public interface RouteService {
-    ArrayList<Route> getRoutes(double latitude, double longitude, double distanceMeters);
+    ArrayList<Route> getRoutes(double latitude, double longitude, int distanceMeters);
 }
